refactor: WorkspaceInviteCard 복사 로직을 공용 훅으로 교체
