+++ frontend/src/shared/utils/getInviteLink/test.ts
@@ -3,27 +3,23 @@ import { describe, expect, it } from "vitest";
 import { getInviteLink } from ".";
 
 describe("getInviteLink", () => {
-  it("참여 코드를 /workspace/code?code= 진입 경로에 넣어 origin이 붙은 링크와 표시용 경로를 만든다", () => {
-    const code = "X35D3S";
+  it("linkToken을 /invite/<linkToken> 진입 경로에 넣어 origin이 붙은 링크와 표시용 경로를 만든다", () => {
+    const linkToken = "Xk3vQ9mZp2LrT7wB1nHc4A";
 
-    const { inviteLink, displayInviteLink } = getInviteLink(code);
+    const { inviteLink, displayInviteLink } = getInviteLink(linkToken);
 
-    expect(displayInviteLink).toBe(`/workspace/code?code=${code}`);
-    expect(inviteLink).toBe(
-      `${window.location.origin}/workspace/code?code=${code}`,
-    );
+    expect(displayInviteLink).toBe(`/invite/${linkToken}`);
+    expect(inviteLink).toBe(`${window.location.origin}/invite/${linkToken}`);
   });
 
   it("URL에 그대로 쓸 수 없는 문자는 인코딩한다", () => {
-    const code = "a/b c?d#e";
+    const linkToken = "a/b c?d#e";
 
-    const { inviteLink, displayInviteLink } = getInviteLink(code);
+    const { inviteLink, displayInviteLink } = getInviteLink(linkToken);
 
-    expect(displayInviteLink).toBe(
-      `/workspace/code?code=${encodeURIComponent(code)}`,
-    );
+    expect(displayInviteLink).toBe(`/invite/${encodeURIComponent(linkToken)}`);
     expect(inviteLink).toBe(
-      `${window.location.origin}/workspace/code?code=${encodeURIComponent(code)}`,
+      `${window.location.origin}/invite/${encodeURIComponent(linkToken)}`,
     );
   });
 });
